+++ src/models/original/SpriteLDA.java
@@ -351,19 +351,21 @@ public class SpriteLDA extends TopicModel implements Serializable {
 		THREAD_DONE_QUEUE = new ArrayBlockingQueue<String>(numThreads);
 		THREADS     = new Worker[numThreads];
 		
-		int dStep = D/numThreads;
-		int zStep = Z/numThreads;
-		int wStep = W/numThreads;
+		float dStep = D/((float)numThreads);
+		float zStep = Z/((float)numThreads);
+		float wStep = W/((float)numThreads);
+		
 		for (int i = 0; i < numThreads; i++) {
-			int minW = wStep*i;
-			int maxW = i < (numThreads-1) ? wStep*(i+1) : W;
-			int minD = dStep*i;
-			int maxD = i < (numThreads-1) ? dStep*(i+1) : D;
-			int minZ = zStep*i;
-			int maxZ = i < (numThreads-1) ? zStep*(i+1) : Z;
+			int minW = (int)(wStep*i);
+			int maxW = i < (numThreads-1) ? (int)(wStep*(i+1)) : W;
+			int minD = (int)(dStep*i);
+			int maxD = i < (numThreads-1) ? (int)(dStep*(i+1)) : D;
+			int minZ = (int)(zStep*i);
+			int maxZ = i < (numThreads-1) ? (int)(zStep*(i+1)) : Z;
 			THREADS[i] = new Worker(minW, maxW, minD, maxD, minZ, maxZ, i);
 			THREADS[i].start();
 		}
+		
 		//}
 	}
